introducing 'lib/pubnub_dns_handler.c/h' and rearragements
